new(demo/drag-ii): refactor to codesandbox example
