+++ packages/vx-demo/src/docs-v2/examples/vx-drag-i/package.json
@@ -19,6 +19,7 @@
     "visualization",
     "d3",
     "react",
-    "vx"
+    "vx",
+    "drag"
   ]
 }
