+++ packages/vx-demo/src/pages/Drag-ii.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import Show from '../components/Show';
-import DragII from '../components/tiles/Drag-ii';
-import DragIISource from '!!raw-loader!../components/tiles/Drag-ii';
+import DragII from '../docs-v2/examples/vx-drag-i/Example';
+import DragIISource from '!!raw-loader!../docs-v2/examples/vx-drag-i/Example';
 
 export default () => {
   return (
-    <Show component={DragII} title="Drag II">
+    <Show component={DragII} title="Drag II" codeSandboxDirectoryName="vx-drag-ii">
       {DragIISource}
     </Show>
   );
